RegAdd Overhaul
Working on issue #5. Removed links from top of the reg_add and reg_order pages. Also changed the workflow - selecting At-Con Registration on the menu now goes to reg_order, which creates a new order array and then redirects to the reg_add page.

Changed cancel button on the reg_add page to clear the current attendee and redirect back to the reg_order page. Cancel on the reg_order page clears the order and redirects back to index.php.